Update logger initialization to allow nil handler
Refined the logger constructor across multiple services and test files to support an optional logging handler parameter. This modification allows a nil value to be passed, which defaults to JSON handler if no specific handler is provided. It enhances flexibility in logging behavior and aids future integration of custom logging mechanisms.

Signed-off-by: SammyOina <sammyoina@gmail.com>